[cef] remove OSR handler
In earlier CEF version special Off-Screen-Handler was required for 
headless mode. Now with ozone frameowrk it is no longer required.
All functionality can be directly move to gui handler
